refactor(slimrpc): split the multicast module and its tests at the ratchet
Both files crossed the 500-line limit with the backpressure fix, and the
repository's answer for that is sub-modules with a thin mod.rs rather than a
recorded exemption.

Both seams are the subject rather than an arbitrary cut. multicast/fanout.rs
holds everything that answers one question — what happens when members
consume at different rates — and nothing else in the module needs to know
the answer. tests/multicast_backpressure.rs holds the three tests that need
an agent chatty enough to overrun a member's channel buffer, which none of
the outcome tests next door do.

No behaviour change: the same nine tests pass, and the binding's fmt,
clippy, build, package and full test gates are green.
