+++ bindings/a2a-protocol-slimrpc/src/multicast/mod.rs
@@ -61,24 +61,11 @@ use crate::error::rpc_error_to_client_error;
 use crate::method;
 use crate::SlimName;
 
+mod fanout;
 mod outcome;
 
+use fanout::{fan_out_to_members, MEMBER_STREAM_CAPACITY};
 pub use outcome::{MemberOutcome, MulticastOutcome};
-
-/// Buffer depth for each member's event stream in a streaming broadcast.
-const MEMBER_STREAM_CAPACITY: usize = 64;
-
-/// How long the fan-out task will wait, once the source is exhausted, to hand a
-/// lagging member the report of what it missed.
-///
-/// A member that falls behind mid-stream is told by the send that carries the
-/// *next* event it receives. One that falls behind on the last events has no
-/// next event, and a stream that simply ends is indistinguishable from one that
-/// ended cleanly — so the tail is flushed explicitly. Bounded, because a
-/// consumer that never reads at all must not hold this task open forever; the
-/// cost of the bound is that such a consumer's stream ends this much later,
-/// which is invisible to a consumer that is not reading.
-const LAG_REPORT_FLUSH: Duration = Duration::from_secs(5);
 
 /// A group of A2A agents addressable with one message.
 pub struct SlimRpcMulticast {
@@ -338,108 +325,8 @@ impl SlimRpcMulticast {
     }
 }
 
-/// Routes each multicast frame to the stream of the member that sent it.
-///
-/// Extracted from [`SlimRpcMulticast::stream_message`] rather than left inline:
-/// the send below is the whole reason this loop is delicate, and inline it sat
-/// under two layers of closure in a function clippy had already grown past its
-/// line limit.
-///
-/// # One consumer's backpressure is not everybody's
-///
-/// Every member has its own channel, which is enough for a member that stays
-/// *silent* — it simply produces no frames. It is not enough for a member whose
-/// *consumer* stops reading, because every member's frames arrive through this
-/// one loop: an awaiting `send` on that member's full channel parks the loop and
-/// with it every other member's stream. Measured 2026-08-19 with one consumer
-/// not reading: another member's live stream reached 151 of 300 events in 25
-/// seconds and never resumed. Non-blocking, it reaches 300 in 220ms.
-///
-/// The trade is the one the SSE fan-out already makes through broadcast's
-/// `Lagged` — a consumer that falls behind loses events rather than holding
-/// everybody else up — and, as there, the gap is *reported* rather than silent.
-/// A task event stream is ordered state, and a consumer that skips from
-/// `Working` to `Completed` cannot otherwise tell that from a task that did.
-async fn fan_out_to_members(
-    channel: Channel,
-    request: pb::SendMessageRequest,
-    timeout: Option<Duration>,
-    metadata: Option<Metadata>,
-    senders: HashMap<String, tokio::sync::mpsc::Sender<ClientResult<StreamResponse>>>,
-) {
-    let frames = channel.multicast_unary_stream::<_, Pb<pb::StreamResponse>>(
-        A2A_SERVICE_NAME,
-        method::SEND_STREAMING_MESSAGE,
-        Pb(request),
-        timeout,
-        metadata,
-    );
-    futures::pin_mut!(frames);
-
-    // Per-member count of events dropped because that member's own channel was
-    // full, delivered to it as an error before the next event it *does*
-    // receive.
-    let mut lagged: HashMap<String, u64> = HashMap::new();
-
-    while let Some(frame) = frames.next().await {
-        let Ok(MulticastItem { context, message }) = frame else {
-            // An unattributable transport error cannot be routed to a member's
-            // stream; the affected member's stream simply ends when this loop
-            // does.
-            continue;
-        };
-        let key = proto_name_key(&context.source);
-        let Some(tx) = senders.get(&key) else {
-            // A frame from an agent that was never invited. Dropping it is
-            // deliberate: the caller asked a specific group, and surfacing a
-            // stranger's events would break that contract.
-            continue;
-        };
-        let event = StreamResponse::try_from(message.into_inner())
-            .map_err(|e| ClientError::Transport(format!("malformed event: {e}")));
-
-        // Tell this member about any earlier gap before handing it the event
-        // that follows the gap. If the report cannot be sent either, the count
-        // survives to the next attempt.
-        if let Some(&missed) = lagged.get(&key).filter(|&&n| n > 0) {
-            if tx.try_send(Err(lag_report(missed))).is_ok() {
-                lagged.remove(&key);
-            }
-        }
-
-        // `try_send`, not `send().await` — see this function's docs. A `Closed`
-        // channel means that member's consumer dropped its stream, which needs
-        // no bookkeeping: the others keep going, because one caller losing
-        // interest in one agent is not a reason to end anybody else's stream.
-        if let Err(tokio::sync::mpsc::error::TrySendError::Full(_)) = tx.try_send(event) {
-            *lagged.entry(key).or_insert(0) += 1;
-        }
-    }
-
-    // Flush what the tail left. See `LAG_REPORT_FLUSH`.
-    for (key, missed) in lagged {
-        if missed == 0 {
-            continue;
-        }
-        let Some(tx) = senders.get(&key) else {
-            continue;
-        };
-        let _ = tx
-            .send_timeout(Err(lag_report(missed)), LAG_REPORT_FLUSH)
-            .await;
-    }
-    // Dropping `senders` ends every member stream that is still open.
-}
-
-/// The error a lagging consumer is handed in place of the events it missed.
-fn lag_report(missed: u64) -> ClientError {
-    ClientError::Transport(format!(
-        "{missed} event(s) dropped: this stream's consumer fell behind"
-    ))
-}
-
 /// The key a [`SlimName`] is matched on when pairing responses to members.
-fn member_key(name: &SlimName) -> String {
+pub(super) fn member_key(name: &SlimName) -> String {
     format!("{}/{}/{}", name.domain, name.namespace, name.service)
 }
 
@@ -455,7 +342,7 @@ fn member_key(name: &SlimName) -> String {
 /// Keying on the full rendering instead is what the first version of this did,
 /// and every response then filed under a name no member matched, so every
 /// member came back as a timeout while all of them had in fact answered.
-fn proto_name_key(source: &ProtoName) -> String {
+pub(super) fn proto_name_key(source: &ProtoName) -> String {
     source
         .to_string()
         .split('/')
