+++ bindings/a2a-protocol-slimrpc/tests/multicast.rs
@@ -335,266 +335,3 @@ async fn inviting_an_agent_that_is_not_on_the_fabric_fails_the_interaction() {
     server.shutdown().await;
     let _ = service.shutdown().await;
 }
-
-// ── One consumer's backpressure must not become everybody's ──────────────────
-//
-// The spec property these three protect is the same one
-// `a_streaming_broadcast_gives_each_agent_its_own_stream` covers for
-// termination: "one agent's stream termination does not affect others". These
-// cover the case that stream cannot reach, because its consumers all read
-// promptly — a consumer that stops reading.
-//
-// Before 2026-08-19, `stream_message` gave each member its own bounded channel
-// and then awaited `send` on it from a single loop shared by every member. One
-// channel per member is enough for a *silent* agent, which produces no frames
-// at all, and does nothing for a slow *consumer*: its channel fills, the shared
-// loop parks, and every other member's stream stops. Measured with one consumer
-// not reading, a live member's stream reached 151 of 300 events in 25 seconds
-// and never resumed. With the non-blocking send it reaches 300 in 220ms.
-
-/// Emits enough events to overrun one member's channel buffer.
-struct Chatty;
-
-const CHATTY_EVENTS: usize = 300;
-
-impl a2a_protocol_server::AgentExecutor for Chatty {
-    fn execute<'a>(
-        &'a self,
-        ctx: &'a a2a_protocol_server::RequestContext,
-        queue: &'a dyn a2a_protocol_server::streaming::EventQueueWriter,
-    ) -> std::pin::Pin<
-        Box<dyn std::future::Future<Output = a2a_protocol_types::error::A2aResult<()>> + Send + 'a>,
-    > {
-        Box::pin(async move {
-            use a2a_protocol_types::events::TaskStatusUpdateEvent;
-            use a2a_protocol_types::task::{ContextId, TaskState, TaskStatus};
-            for _ in 0..CHATTY_EVENTS {
-                queue
-                    .write(a2a_protocol_types::StreamResponse::StatusUpdate(
-                        TaskStatusUpdateEvent {
-                            task_id: ctx.task_id.clone(),
-                            context_id: ContextId::new(ctx.context_id.clone()),
-                            status: TaskStatus::new(TaskState::Working),
-                            metadata: None,
-                        },
-                    ))
-                    .await?;
-            }
-            queue
-                .write(a2a_protocol_types::StreamResponse::StatusUpdate(
-                    TaskStatusUpdateEvent {
-                        task_id: ctx.task_id.clone(),
-                        context_id: ContextId::new(ctx.context_id.clone()),
-                        status: TaskStatus::with_timestamp(TaskState::Completed),
-                        metadata: None,
-                    },
-                ))
-                .await?;
-            Ok(())
-        })
-    }
-}
-
-/// A two-agent group whose agents each emit [`CHATTY_EVENTS`] events.
-struct ChattyGroup {
-    service: Arc<Service>,
-    servers: Vec<Arc<SlimRpcServer>>,
-    multicast: SlimRpcMulticast,
-}
-
-impl ChattyGroup {
-    async fn new(test_name: &str) -> Self {
-        let service = common::service(test_name);
-        let invited: Vec<SlimName> = (0..2)
-            .map(|i| SlimName::new("org", "mc", format!("agent-{i}")))
-            .collect();
-
-        let mut servers = Vec::new();
-        for name in &invited {
-            let parts = common::app_for(&service, name, &name.service);
-            let handler = Arc::new(
-                a2a_protocol_server::RequestHandlerBuilder::new(Chatty)
-                    .with_agent_card(common::agent_card(name))
-                    .allow_unauthenticated_extended_card()
-                    .build()
-                    .expect("build handler"),
-            );
-            servers.push(Arc::new(SlimRpcServer::from_app(
-                parts,
-                handler,
-                name.clone(),
-            )));
-        }
-
-        let caller = SlimName::new("org", "mc", "caller");
-        let (caller_app, _) = common::app_for(&service, &caller, "caller");
-        let multicast = SlimRpcMulticast::from_app(caller_app, invited)
-            .expect("open the group channel")
-            .with_timeout(Duration::from_secs(20));
-
-        for server in &servers {
-            let s = Arc::clone(server);
-            tokio::spawn(async move {
-                let _ = s.serve().await;
-            });
-        }
-        tokio::time::sleep(Duration::from_millis(200)).await;
-
-        Self {
-            service,
-            servers,
-            multicast,
-        }
-    }
-
-    async fn shutdown(self) {
-        for server in &self.servers {
-            server.shutdown().await;
-        }
-        let _ = self.service.shutdown().await;
-    }
-}
-
-/// A consumer that stops reading must not stop the other members' streams.
-///
-/// The stalled stream is *held*, not dropped: a dropped stream closes its
-/// channel, which the old code already handled. What it could not handle was a
-/// consumer that is still there and simply not reading — the state a task
-/// switched away from, or one blocked on something else, is in.
-#[tokio::test]
-async fn a_consumer_that_stops_reading_does_not_stall_the_other_members() {
-    let group = ChattyGroup::new("mc-stall").await;
-
-    let mut streams = group
-        .multicast
-        .stream_message(common::send_params("stall one of us"), None)
-        .expect("opening the broadcast must succeed");
-    assert_eq!(streams.len(), 2);
-
-    let (_stalled_name, stalled) = streams.remove(0);
-    let (live_name, mut live) = streams.remove(0);
-
-    let mut seen = 0usize;
-    let finished = tokio::time::timeout(Duration::from_secs(20), async {
-        while let Some(event) = live.next().await {
-            if event.is_ok() {
-                seen += 1;
-            }
-            if seen >= CHATTY_EVENTS {
-                return true;
-            }
-        }
-        false
-    })
-    .await;
-
-    assert_eq!(
-        finished,
-        Ok(true),
-        "{}'s stream must complete while another member's consumer is not reading; \
-         it saw {seen} of {CHATTY_EVENTS} events",
-        live_name.service
-    );
-
-    drop(stalled);
-    group.shutdown().await;
-}
-
-/// The stalled consumer is *told* it missed events, rather than silently
-/// handed a stream with a hole in it.
-///
-/// This is the half that makes dropping acceptable. A task event stream is
-/// ordered state: a consumer that silently skips from `Working` to `Completed`
-/// cannot tell that from a task that did exactly that.
-#[tokio::test]
-async fn a_consumer_that_falls_behind_is_told_what_it_missed() {
-    let group = ChattyGroup::new("mc-lag").await;
-
-    let mut streams = group
-        .multicast
-        .stream_message(common::send_params("outrun one of us"), None)
-        .expect("opening the broadcast must succeed");
-
-    let (_a_name, mut a) = streams.remove(0);
-    let (_b_name, mut b) = streams.remove(0);
-
-    // Let one member's channel overrun while nothing reads it.
-    tokio::time::sleep(Duration::from_secs(2)).await;
-
-    // Then drain it and look for the report.
-    let mut saw_lag_report = false;
-    for _ in 0..CHATTY_EVENTS {
-        match tokio::time::timeout(Duration::from_secs(5), a.next()).await {
-            Ok(Some(Err(e))) => {
-                let msg = e.to_string();
-                if msg.contains("dropped") && msg.contains("fell behind") {
-                    saw_lag_report = true;
-                    break;
-                }
-            }
-            Ok(Some(Ok(_))) => {}
-            Ok(None) | Err(_) => break,
-        }
-    }
-
-    assert!(
-        saw_lag_report,
-        "a consumer that fell behind must be told the stream has a gap, not \
-         silently handed the events that survived"
-    );
-
-    // Drain the other one so the shutdown below is not racing a live stream.
-    while let Ok(Some(_)) = tokio::time::timeout(Duration::from_millis(200), b.next()).await {}
-    group.shutdown().await;
-}
-
-/// The control: when every consumer reads, nobody loses anything.
-///
-/// Without this the two tests above pass for a `stream_message` that drops
-/// every event it cannot send instantly, which would be a worse binding than
-/// the one being fixed.
-#[tokio::test]
-async fn every_member_gets_every_event_when_its_consumer_keeps_up() {
-    let group = ChattyGroup::new("mc-keepup").await;
-
-    let streams = group
-        .multicast
-        .stream_message(common::send_params("everyone reads"), None)
-        .expect("opening the broadcast must succeed");
-    assert_eq!(streams.len(), 2);
-
-    let readers: Vec<_> = streams
-        .into_iter()
-        .map(|(member, mut stream)| {
-            tokio::spawn(async move {
-                let mut ok = 0usize;
-                let mut errors = Vec::new();
-                while let Ok(Some(event)) =
-                    tokio::time::timeout(Duration::from_secs(20), stream.next()).await
-                {
-                    match event {
-                        Ok(_) => ok += 1,
-                        Err(e) => errors.push(e.to_string()),
-                    }
-                }
-                (member, ok, errors)
-            })
-        })
-        .collect();
-
-    for reader in readers {
-        let (member, ok, errors) = reader.await.expect("reader task");
-        assert!(
-            errors.is_empty(),
-            "{} saw errors on a stream nothing was lagging: {errors:?}",
-            member.service
-        );
-        assert!(
-            ok >= CHATTY_EVENTS,
-            "{} saw {ok} events, expected at least {CHATTY_EVENTS}",
-            member.service
-        );
-    }
-
-    group.shutdown().await;
-}
